Camera Rotation
Startet with the Camera setup. Camera is now in a different position directly behind the player and the rotation to the left and right is restricted to 45°.
TODO: Zoom in and Zoom out implementation when the player uses up and down on the right stick. It does nothing at the moment.