Changes in Camera and Movement
Changed the character position and the camera position. Is hopefully easier for collition this way. Changed also the feeling of the movement. It is a little harder to move and the movement feels a little more realistic. Falling in Air has resistance for movement. Also fixed the problem with the character moving without input.